Move `gen_rule_consts` to the codegen module
